hash table works but in a negative way...QAQ

+++ statusBoard.h
@@ -16,10 +16,10 @@ typedef struct {
 
     //followings are arrays for DIRECT grade estimate
     //0 for black and 1 for white orz
-    int row_grade[2][BOARD_SIZE];
-    int col_grade[2][BOARD_SIZE];
-    int oblique_sum_grade[2][2 * BOARD_SIZE + 5];
-    int oblique_delta_grade[2][2 * BOARD_SIZE + 5];
+    long long row_grade[2][BOARD_SIZE];
+    long long col_grade[2][BOARD_SIZE];
+    long long oblique_sum_grade[2][2 * BOARD_SIZE + 5];
+    long long oblique_delta_grade[2][2 * BOARD_SIZE + 5];
     long long total_grade[2];
 
     //these are used to record the last place and the total steps
